Only compiles qt/opengl source files if QT and OpenGL present
Modified the CMakeLists.txt files so that source files that depend on QT and OpenGL are not compiled if QT and OpenGl are
not found on the system. Uses the CheckQTOpenGL flag to determine whether QT and OpenGL are available.

Tested and now compiles without error. Runtime test still needs to be made using torque on the Alliance wheeler cluster.